tighten up abilityBot with runTurn
